% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deframe_with_rownames.R
\name{deframe_with_rownames}
\alias{deframe_with_rownames}
\title{Convert a one column data frame with rownames to a list}
\usage{
deframe_with_rownames(.x)
}
\arguments{
\item{.x}{A data frame}
}
\value{
A list with rownames as the list names, and the values as the data frame column.
}
\description{
Convert a one column data frame with rownames to a list
}
\details{
The \code{\link[tibble:enframe]{tibble::deframe()}} function is awesome, but since rownames are always
special, it doesn't quite work as I'd like. This function will transform a single
column data frame (with rownames) into a named list. It appears that \code{\link[base:drop]{base::drop()}}
does not honor row names so this seems the best option.
}
\note{
The data frame column name is lost.
}
\examples{
\dontrun{
# Iris has numbers for rownames, so this will have number=value
deframe_with_rownames(data.frame(iris[,1]))
}
}
