% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in.R
\name{\%not in\%}
\alias{\%not in\%}
\title{Negative Value Matching}
\usage{
x \%not in\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched. See [\code{\link[base:match]{base::match()}}] for details.}

\item{table}{vector or NULL: the values to be matched against. See [\code{\link[base:match]{base::match()}}] for details.}
}
\value{
A vector of the same length as \code{x}.
}
\description{
Returns a logical vector indicating if there is \emph{not} a match for its left operand.
}
\details{
This is a convenience function that negates the normal [\code{\link[base:match]{base::\%in\%()}}] function.
}
\examples{
\dontrun{
1 \%not in\% c(1,2,3)
}
}
\seealso{
base::match()
}
