% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or.R
\name{\%or\%}
\alias{\%or\%}
\title{Extract value or return default}
\usage{
a \%or\% b
}
\arguments{
\item{a}{Value to use, unless empty.}

\item{b}{Alternative value to use, if a is empty.}
}
\value{
a unless empty, in which case b is returned.
}
\description{
A pretty clever approach for accessing part of a list,
returning a default if there is nothing of that name in the list.
This was taken from https://stackoverflow.com/questions/26561518/built-in-for-getting-list-element-or-default
}
\examples{

NULL \%or\% "default"

}
