% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.not_na}
\alias{is.not_na}
\title{Test if input elements are \emph{not} missing}
\usage{
is.not_na(x)
}
\arguments{
\item{x}{A value or vector of values (see [\code{\link[base:NA]{base::is.na()}}]).}
}
\value{
Logical indicating if the values are \emph{not} NA (see [\code{\link[base:NA]{base::is.na()}}]).
}
\description{
Performs the inverse of the [\code{\link[base:NA]{base::is.na()}}] function (syntactic sugar).
}
\details{
This is not a very exciting function, it merely packages the logic of

\if{html}{\out{<div class="sourceCode">}}\preformatted{!is.na(x)
}\if{html}{\out{</div>}}

in a slightly more readable format.
}
\examples{
is.not_na(c(0,1,NA))

}
